Refactor the plugin Fixes #8 and fixes #9
By not using the filetype as the key, we can have duplicate supported filetypes. This makes adding new formatters as painless as possible.
I've also completely revamped the list command to show as a table in the log, with the current on/off status.

PS: the 'fmt' command now runs the formatter, and 'fmt list' is to show the available formatters.